simplify bucket prefix
